move iterators back to `return_enums.rs`
